# Conflicts:
#	database/migrations/2023_04_21_161603_create_roads_table.php
#	database/migrations/2023_04_23_050137_create_plot_allocations_table.php
#	resources/views/vendor/backpack/base/inc/sidebar_content.blade.php
#	routes/backpack/custom.php

+++ database/migrations/2023_04_01_170553_create_members_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->text('present_address')->nullable();
             $table->string('mobile', 20)->nullable();
             $table->string('phone', 20)->nullable();
-            $table->integer('national_id')->nullable();
+            $table->bigInteger('national_id')->nullable();
             $table->foreignId('occupation_id')->nullable()->constrained();
             $table->foreignId('marital_status')->nullable()->constrained('marital_statuses');
             $table->timestamps();
